third larger attempt to split up files and design the yarn properly
split up files accordingly into the YarnLine on its own, Contactables as Behaviours that attach to each type of object, and Contacts which are created and destroyed to maintain the line.

code is not working yet. need to fix pushpin errors and YarnLine does not yet appropriately update its Contacts